Improve linear algebra

+++ src/main/java/com/meow/bebrablender/math/affine/AffineApplicable.java
@@ -1,7 +1,7 @@
 package com.meow.bebrablender.math.affine;
 
-import com.meow.bebrablender.math.vectors.Vector3f;
+import com.meow.bebrablender.math.vectors.Vector3d;
 
 public interface AffineApplicable {
-    Vector3f apply(Vector3f v);
+    Vector3d apply(Vector3d v);
 }
