+++ src/test/java/com/meow/bebrablender/math/affine/AffineApplicableTest.java
@@ -1,20 +1,20 @@
 package com.meow.bebrablender.math.affine;
 
-import com.meow.bebrablender.math.vectors.Vector3f;
+import com.meow.bebrablender.math.vectors.Vector3d;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 class AffineApplicableTest {
     @Test
     void rotateX45() {
-        Vector3f v = new Vector3f(4, 5, 9);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(4, 5, 9);
+        Vector3d c = v.copy();
 
         double a = 45;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateX(a).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateX(a).apply(v);
+        Vector3d actual = new Vector3d(
                 c.x(),
                 c.y() * Math.cos(ra) + Math.sin(ra) * c.z(),
                 Math.cos(ra) * c.z() - Math.sin(ra) * c.y()
@@ -25,14 +25,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateX30() {
-        Vector3f v = new Vector3f(1, 6, 2);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 6, 2);
+        Vector3d c = v.copy();
 
         double a = 30;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateX(a).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateX(a).apply(v);
+        Vector3d actual = new Vector3d(
                 c.x(),
                 c.y() * Math.cos(ra) + Math.sin(ra) * c.z(),
                 -Math.sin(ra) * c.y() + Math.cos(ra) * c.z()
@@ -43,15 +43,15 @@ class AffineApplicableTest {
 
     @Test
     void rotateX0() {
-        Vector3f v = new Vector3f(1, 0, 1);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 0, 1);
+        Vector3d c = v.copy();
 
         double a = 0;
         double ra = Math.toRadians(a);
 
 
-        Vector3f expected = new RotateX(a).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateX(a).apply(v);
+        Vector3d actual = new Vector3d(
                 c.x(),
                 c.y() * Math.cos(ra) + Math.sin(ra) * c.z(),
                 -Math.sin(ra) * c.y() + Math.cos(ra) * c.z()
@@ -62,14 +62,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateY45() {
-        Vector3f v = new Vector3f(4, 5, 9);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(4, 5, 9);
+        Vector3d c = v.copy();
 
         double a = 45;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateY(a).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateY(a).apply(v);
+        Vector3d actual = new Vector3d(
                 Math.cos(ra) * c.x() + Math.sin(ra) * c.z(),
                 c.y(),
                 -Math.sin(ra) * c.x() + Math.cos(ra) * c.z()
@@ -80,14 +80,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateY30() {
-        Vector3f v = new Vector3f(1, 6, 2);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 6, 2);
+        Vector3d c = v.copy();
 
         double a = 30;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateY(a).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateY(a).apply(v);
+        Vector3d actual = new Vector3d(
                 Math.cos(ra) * c.x() + Math.sin(ra) * c.z(),
                 c.y(),
                 -Math.sin(ra) * c.x() + Math.cos(ra) * c.z()
@@ -98,14 +98,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateY0() {
-        Vector3f v = new Vector3f(1, 0, 1);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 0, 1);
+        Vector3d c = v.copy();
 
         double a = 0;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateY(ra).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateY(ra).apply(v);
+        Vector3d actual = new Vector3d(
                 Math.cos(ra) * c.x() + Math.sin(ra) * c.z(),
                 c.y(),
                 -Math.sin(ra) * c.x() + Math.cos(ra) * c.z()
@@ -116,14 +116,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateZ45() {
-        Vector3f v = new Vector3f(4, 5, 9);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(4, 5, 9);
+        Vector3d c = v.copy();
 
         double a = 45;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateZ(ra).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateZ(ra).apply(v);
+        Vector3d actual = new Vector3d(
                 c.x() * Math.cos(ra) + Math.sin(ra) * c.y(),
                 -Math.sin(ra) * c.x() + Math.cos(ra) * c.y(),
                 c.z()
@@ -134,14 +134,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateZ30() {
-        Vector3f v = new Vector3f(1, 6, 2);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 6, 2);
+        Vector3d c = v.copy();
 
         double a = 30;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateZ(ra).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateZ(ra).apply(v);
+        Vector3d actual = new Vector3d(
                 c.x() * Math.cos(ra) + Math.sin(ra) * c.y(),
                 -Math.sin(ra) * c.x() + Math.cos(ra) * c.y(),
                 c.z()
@@ -152,14 +152,14 @@ class AffineApplicableTest {
 
     @Test
     void rotateZ0() {
-        Vector3f v = new Vector3f(1, 0, 1);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 0, 1);
+        Vector3d c = v.copy();
 
         double a = 0;
         double ra = Math.toRadians(a);
 
-        Vector3f expected = new RotateZ(ra).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new RotateZ(ra).apply(v);
+        Vector3d actual = new Vector3d(
                 c.x() * Math.cos(ra) + Math.sin(ra) * c.y(),
                 -Math.sin(ra) * c.x() + Math.cos(ra) * c.y(),
                 c.z()
@@ -171,15 +171,15 @@ class AffineApplicableTest {
 
     @Test
     void scale() {
-        Vector3f v = new Vector3f(2, 6, 8);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(2, 6, 8);
+        Vector3d c = v.copy();
         
         double sX = 8;
         double sY = 4;
         double sZ = 2;
 
-        Vector3f expected = new Scale(sX, sY, sZ).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new Scale(sX, sY, sZ).apply(v);
+        Vector3d actual = new Vector3d(
                 sX * c.x(),
                 sY * c.y(),
                 sZ * c.z()
@@ -190,15 +190,15 @@ class AffineApplicableTest {
 
     @Test
     void translate() {
-        Vector3f v = new Vector3f(1, 6, 3);
-        Vector3f c = v.copy();
+        Vector3d v = new Vector3d(1, 6, 3);
+        Vector3d c = v.copy();
         
         double sX = 2;
         double sY = 10;
         double sZ = 8;
 
-        Vector3f expected = new Scale(sX, sY, sZ).apply(v);
-        Vector3f actual = new Vector3f(
+        Vector3d expected = new Scale(sX, sY, sZ).apply(v);
+        Vector3d actual = new Vector3d(
                 sX + c.x(),
                 sY + c.y(),
                 sZ + c.z()
